Object rotation now as enum

+++ src/Game/Objects/ITank.cpp
@@ -3,7 +3,7 @@
 ITank::ITank(ObjectType type, const sf::Vector2f& position) : IObject(type, position, sf::Vector2f(16, 16)) {}
 
 uint8_t ITank::getState() const {
-  return _wheelState << 2 | _rotation;
+  return _wheelState << 2 | (uint8_t)_rotation;
 }
 
 void ITank::snap(float& coordinate) {
